+++ plugins/flutter/android/src/main/kotlin/za/co/synthesis/halo/sdkflutterplugin/UIContext.kt
@@ -1,0 +1,20 @@
+package za.co.synthesis.halo.sdkflutterplugin
+
+import android.app.Activity
+import java.lang.ref.WeakReference
+
+object UIContext {
+    private var activityRef: WeakReference<Activity>? = null
+    private var enableSchemeAnimations: Boolean = false
+
+    fun updateActivity(activity: Activity?) {
+        activityRef = activity?.let { WeakReference(it) }
+    }
+
+    fun enableSchemeAnimations(enable: Boolean){
+        enableSchemeAnimations = enable;
+    }
+
+    fun getActivity(): Activity? = activityRef?.get()
+    fun areSchemeAnimationsEnabled(): Boolean = enableSchemeAnimations
+}